requirements.txt
# Core Dependencies
# ================

# OpenAI API client for GPT models (code generation, optimization, chat)
openai>=1.0.0

# Numerical computing for embeddings and vector operations
numpy>=1.24.0

# Machine learning utilities (cosine similarity for vector search)
scikit-learn>=1.3.0

# Environment variable management for API keys
python-dotenv>=1.0.0

# Code Analysis & Graph Generation
# =================================

# Python's built-in Abstract Syntax Tree parser (included in Python standard library)
# ast - no installation needed

# Static Control Flow Graph generation
staticfg>=0.1.0

# Graph visualization library
graphviz>=0.20.0

# Alternative CFG library (uncomment if preferred over staticfg)
# pycfg>=0.1.0

# Web Framework & API
# ===================

# Fast API framework for building REST APIs
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Frontend framework for building interactive web apps
streamlit>=1.28.0

# PDF Processing for Part B
# ==========================

# PDF reading and text extraction
pypdf2>=3.0.0

# PDF creation from text
reportlab>=4.0.0

# Additional Utilities
# ====================

# HTTP library for API requests
requests>=2.31.0

# Data manipulation and analysis
pandas>=2.0.0

# JSON schema validation
pydantic>=2.0.0

# Pretty printing and formatting
rich>=13.0.0

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
